Clean up some warnings on Windows builds
* src/arscan.c (parse_int): Avoid conversion from int to char and
check for overflow given a max value.
(ar_scan): Check intmax sizes then cast to the proper type.
(ar_member_touch): Get proper return type from ar_scan and cast it
to off_t.
* src/function.c (a_word_hash_cmp): Don't cast from size_t to int.
(func_filter_filterout): Count using unsigned long to avoid casts.
(construct_include_path): Explicitly cast to void*.
* src/shuffle.c (reverse_shuffle_array): Use size_t index.
* src/w32/subproc/sub_proc.c (process_wait_for_multiple_objects):
Initialize the return value in case the count is 0.

+++ src/shuffle.c
@@ -128,7 +128,7 @@ reverse_shuffle_array (void **a, size_t len)
       void *t;
 
       /* Pick mirror and swap. */
-      unsigned int j = len - 1 - i;
+      size_t j = len - 1 - i;
 
       /* Swap. */
       t = a[i];
